requirement.txt
asgiref==3.5.0
Django==4.0.3
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-extensions==3.1.5
django-filter==22.1
django-leaflet==0.28.3
djangorestframework==3.13.1
djangorestframework-gis==1.0
djangorestframework-jsonapi==6.0.0
inflection==0.5.1
pkg_resources==0.0.0
psycopg2-binary==2.9.3
pytz==2022.1
sqlparse==0.4.3
typing_extensions==4.1.1
